+++ tests/Nelmio/Alice/Fixtures/LoaderTest.php
@@ -949,6 +949,32 @@ class LoaderTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('bar', $res['user_bar']->username);
     }
 
+    /**
+     * @group legacy
+     */
+    public function testLoadCreatesEnumsOfObjectsWithMalformedList()
+    {
+        $res = $this->loadData([
+            self::USER => [
+                'user_{alice, bob, foo bar}' => [
+                    'username' => '<current()>',
+                    'email' => '<current()>@gmail.com',
+                ],
+            ],
+        ]);
+
+        $this->assertCount(3, $res);
+
+        $this->assertInstanceOf(self::USER, $res['user_alice']);
+        $this->assertEquals('alice', $res['user_alice']->username);
+
+        $this->assertInstanceOf(self::USER, $res['user_bob']);
+        $this->assertEquals('bob', $res['user_bob']->username);
+
+        $this->assertInstanceOf(self::USER, $res['user_foo bar']);
+        $this->assertEquals('foo bar', $res['user_foo bar']->username);
+    }
+
     public function testLocalObjectsAreNotReturned()
     {
         $res = $this->loadData([
@@ -1005,6 +1031,35 @@ class LoaderTest extends \PHPUnit_Framework_TestCase
                     'favoriteNumber' => 2
                 ],
                 '{user, uzer}3 (extends user2)' => [
+                    'fullname' => 'testfullname'
+                ],
+            ],
+        ]);
+
+        $this->assertCount(2, $res);
+        foreach (['user3', 'uzer3'] as $key) {
+            $this->assertInstanceOf(self::USER, $this->loader->getReference($key));
+
+            $this->assertSame($this->loader->getReference($key)->email, 'base@email.com');
+            $this->assertSame($this->loader->getReference($key)->favoriteNumber, 2);
+            $this->assertSame($this->loader->getReference($key)->fullname, 'testfullname');
+        }
+    }
+
+    /**
+     * @group legacy
+     */
+    public function testTemplateCanExtendOtherTemplateObjectsCombinedWithRangeWithLegacySyntax()
+    {
+        $res = $this->loadData([
+            self::USER => [
+                'us{er,rr} (template)' => [
+                    'email' => 'base@email.com'
+                ],
+                'user{1..2} (template, extends user)' => [
+                    'favoriteNumber' => 2
+                ],
+                '{user,uzer}3 (extends user2)' => [
                     'fullname' => 'testfullname'
                 ],
             ],